+++ src/components/SendFlow/SuccessStep.tsx
@@ -12,7 +12,7 @@ import {
 import { Link } from "react-router-dom";
 
 import React, { useContext } from "react";
-import { getHashUrl } from "../../views/operations/operationsUtils";
+import { getHashUrl } from "../../views/operations/utils";
 import { TzktLink } from "../TzktLink";
 import { DynamicModalContext } from "../DynamicModal";
 import { useSelectedNetwork } from "../../utils/hooks/networkHooks";
