--- src/views/operations/operationsUtils.ts
@@ -1,278 +1,0 @@
-import { formatRelative } from "date-fns";
-import { z } from "zod";
-import { tokenPrettyAmount } from "../../types/Token";
-import { OperationDisplay, TezTransfer, TokenTransfer } from "../../types/Transfer";
-import { fromRaw } from "../../types/Token";
-import { compact } from "lodash";
-import { getIPFSurl } from "../../utils/token/nftUtils";
-import { BigNumber } from "bignumber.js";
-import { prettyTezAmount } from "../../utils/format";
-import { DelegationOperation } from "@tzkt/sdk-api";
-import { parsePkh } from "../../types/Address";
-import { Network } from "../../types/Network";
-
-// When submitting batch it makes sense to use the whole batch URL which can be accessed by the operation hash
-// But when it comes to showing individual operations, it's worth adding nonce counter to the URL to navigate to the exact operation
-export const getHashUrl = ({
-  hash,
-  counter,
-  network,
-}: {
-  hash: string;
-  counter?: number;
-  network: Network;
-}) => compact([network.tzktExplorerUrl, hash, counter]).join("/");
-
-export const getTransactionUrl = ({
-  transactionId,
-  originationId,
-  migrationId,
-  network,
-}: {
-  transactionId: number | undefined;
-  originationId: number | undefined;
-  migrationId: number | undefined;
-  network: Network;
-}) => {
-  if (transactionId) {
-    return `${network.tzktExplorerUrl}/transactions/${transactionId}`;
-  }
-  if (originationId) {
-    return `${network.tzktExplorerUrl}/originations/${originationId}`;
-  }
-  if (migrationId) {
-    return `${network.tzktExplorerUrl}/migrations/${migrationId}`;
-  }
-  throw new Error("Cannot find transaction TzKT URL");
-};
-
-export const getIsInbound = (prettyAmount: string) => prettyAmount[0] === "+";
-
-export const getAmountColor = (prettyAmount: string) => {
-  const sign = prettyAmount[0];
-  if (sign === "+") {
-    return "umami.green";
-  } else if (sign === "-") {
-    return "umami.orange";
-  }
-  return "umami.gray.400";
-};
-
-const getSign = (address: string, sender: string, recipient: string) => {
-  if (address === sender) {
-    return "-";
-  }
-
-  if (address === recipient) {
-    return "+";
-  }
-
-  throw new Error(`Address ${address} doesn't match sender or recipient`);
-};
-
-export const getKey = (op: OperationDisplay) => {
-  // id is unique, but it might be that we have
-  // both the recipient and the sender accounts
-  // and in that case the only way to differentiate
-  // would be to use the operation sign (+/-)
-  return op.amount.prettyDisplay + op.id;
-};
-
-const Address = { address: z.string() };
-
-const TezTransaction = z.object({
-  id: z.number(),
-  type: z.string(),
-  sender: z.object(Address),
-  target: z.object(Address),
-  timestamp: z.string(),
-  amount: z.number(),
-  hash: z.string(),
-  level: z.number(),
-});
-
-export const getTezOperationDisplay = (
-  transfer: TezTransfer,
-  forAddress: string,
-  network: Network
-) => {
-  const parseResult = TezTransaction.safeParse(transfer);
-  if (!parseResult.success) {
-    console.warn("getTezOperationDisplay failed parsing");
-    return null;
-  }
-  const parsed = parseResult.data;
-  const sender = parsed.sender.address;
-  const target = parsed.target.address;
-
-  const sign = getSign(forAddress, sender, target);
-
-  const prettyTimestamp = formatRelative(new Date(parsed.timestamp), new Date());
-
-  const result: OperationDisplay = {
-    id: parsed.id,
-    amount: {
-      prettyDisplay: sign + prettyTezAmount(new BigNumber(parsed.amount)),
-    },
-    prettyTimestamp,
-    timestamp: parsed.timestamp,
-    recipient: parsePkh(parsed.target.address),
-    sender: parsePkh(parsed.sender.address),
-    type: "transaction",
-    tzktUrl: getHashUrl({ ...parsed, network }),
-    fee:
-      transfer.bakerFee !== undefined
-        ? prettyTezAmount(new BigNumber(transfer.bakerFee))
-        : undefined,
-    status: "confirmed",
-    level: parsed.level,
-  };
-  return result;
-};
-
-const TokenTransaction = z.object({
-  id: z.number(),
-  // When the "from" field is missing, we assume that the token is minted by the contract.
-  from: z.object(Address).optional(),
-  to: z.object(Address),
-  token: z.object({
-    contract: z.object(Address),
-  }),
-  timestamp: z.string(),
-  amount: z.string(),
-  level: z.number(),
-  transactionId: z.number().optional(),
-  migrationId: z.number().optional(),
-  originationId: z.number().optional(),
-});
-
-export const getTokenOperationDisplay = (
-  transfer: TokenTransfer,
-  forAddress: string,
-  network: Network
-) => {
-  const token = fromRaw(transfer.token);
-
-  const transferRequired = TokenTransaction.safeParse(transfer);
-
-  if (!token || !transferRequired.success) {
-    console.warn("getTokenOperationDisplay failed parsing");
-    return null;
-  }
-
-  const parsed = transferRequired.data;
-
-  const sender = parsed.from?.address || parsed.token.contract.address;
-
-  const metadata = transfer.token.metadata;
-
-  const sign = getSign(forAddress, sender, parsed.to.address);
-
-  const uri = metadata && (metadata.thumbnailUri || metadata.displayUri);
-  const displayId = transfer.token.id;
-
-  const level = parsed.level;
-
-  const prettyTimestamp = formatRelative(new Date(parsed.timestamp), new Date());
-
-  let prettyAmount: string;
-  if (token.type === "nft") {
-    prettyAmount = transfer.amount;
-  } else {
-    prettyAmount = tokenPrettyAmount(transfer.amount, token, { showSymbol: true });
-  }
-
-  const result: OperationDisplay = {
-    id: parsed.id,
-    type: "transaction",
-    amount: {
-      prettyDisplay: sign + prettyAmount,
-      url: uri && getIPFSurl(uri),
-      id: displayId,
-    },
-    prettyTimestamp,
-    timestamp: parsed.timestamp,
-    recipient: parsePkh(parsed.to.address),
-    sender: parsePkh(sender),
-    tzktUrl: getTransactionUrl({
-      transactionId: parsed.transactionId,
-      originationId: parsed.originationId,
-      migrationId: parsed.migrationId,
-      network,
-    }),
-    level,
-  };
-  return result;
-};
-
-const DelegationSchema = z.object({
-  id: z.number(),
-  sender: z.object(Address),
-  newDelegate: z.object(Address).optional(),
-  timestamp: z.string(),
-  amount: z.number(),
-  hash: z.string(),
-  level: z.number(),
-  bakerFee: z.number(),
-});
-
-const getDelegationOperationDisplay = (
-  delegation: DelegationOperation,
-  network: Network
-): OperationDisplay | null => {
-  const parseResult = DelegationSchema.safeParse(delegation);
-
-  if (!parseResult.success) {
-    console.warn("getDelegationOperationDisplay failed parsing", parseResult.error);
-
-    return null;
-  }
-
-  const parsed = parseResult.data;
-
-  if (!parsed.newDelegate) {
-    return null;
-  }
-
-  const prettyAmount = prettyTezAmount(new BigNumber(parsed.amount));
-
-  const prettyTimestamp = formatRelative(new Date(parsed.timestamp), new Date());
-  const level = parsed.level;
-
-  return {
-    id: parsed.id,
-    type: "delegation",
-    amount: {
-      prettyDisplay: prettyAmount,
-    },
-    prettyTimestamp,
-    timestamp: parsed.timestamp,
-    recipient: parsePkh(parsed.newDelegate.address),
-    sender: parsePkh(parsed.sender.address),
-    tzktUrl: getHashUrl({ ...parsed, network }),
-    level,
-    fee: prettyTezAmount(new BigNumber(parsed.bakerFee)),
-  };
-};
-
-export const sortOperationsByTimestamp = (ops: OperationDisplay[]) => {
-  return [...ops].sort((a, b) => {
-    return new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime();
-  });
-};
-
-export const getOperationDisplays = (
-  tezTransfers: TezTransfer[] = [],
-  tokenTransfers: TokenTransfer[] = [],
-  delegation: DelegationOperation | null = null,
-  forAdress: string,
-  network: Network
-) => {
-  return sortOperationsByTimestamp(
-    compact([
-      ...tezTransfers.map(t => getTezOperationDisplay(t, forAdress, network)),
-      ...tokenTransfers.map(t => getTokenOperationDisplay(t, forAdress, network)),
-      delegation ? getDelegationOperationDisplay(delegation, network) : null,
-    ])
-  );
-};
